Squash BlockMetadataStore and BlockDagKeyValueStorage
